Move Chinese Wiktionary extract pronunciation code to a new file
